Refactor to avoid duplicating default_libcall_names
