+++ cranelift-faerie/src/backend.rs
@@ -46,7 +46,7 @@ impl FaerieBuilder {
     /// The `libcall_names` function provides a way to translate `cranelift_codegen`'s `ir::LibCall`
     /// enum to symbols. LibCalls are inserted in the IR as part of the legalization for certain
     /// floating point instructions, and for stack probes. If you don't know what to use for this
-    /// argument, use `FaerieBuilder::default_libcall_names()`.
+    /// argument, use `cranelift_module::default_libcall_names()`.
     pub fn new(
         isa: Box<TargetIsa>,
         name: String,
@@ -63,25 +63,6 @@ impl FaerieBuilder {
             name,
             collect_traps,
             libcall_names,
-        })
-    }
-
-    /// Default names for `ir::LibCall`s. A function by this name is imported into the object as
-    /// part of the translation of a `ir::ExternalName::LibCall` variant.
-    pub fn default_libcall_names() -> Box<Fn(ir::LibCall) -> String> {
-        Box::new(move |libcall| match libcall {
-            ir::LibCall::Probestack => "__cranelift_probestack".to_owned(),
-            ir::LibCall::CeilF32 => "ceilf".to_owned(),
-            ir::LibCall::CeilF64 => "ceil".to_owned(),
-            ir::LibCall::FloorF32 => "floorf".to_owned(),
-            ir::LibCall::FloorF64 => "floor".to_owned(),
-            ir::LibCall::TruncF32 => "truncf".to_owned(),
-            ir::LibCall::TruncF64 => "trunc".to_owned(),
-            ir::LibCall::NearestF32 => "nearbyintf".to_owned(),
-            ir::LibCall::NearestF64 => "nearbyint".to_owned(),
-            ir::LibCall::Memcpy => "memcpy".to_owned(),
-            ir::LibCall::Memset => "memset".to_owned(),
-            ir::LibCall::Memmove => "memmove".to_owned(),
         })
     }
 }
