+++ cranelift-simplejit/examples/simplejit-minimal.rs
@@ -1,12 +1,11 @@
 use cranelift::prelude::*;
-use cranelift_module::{Linkage, Module};
+use cranelift_module::{default_libcall_names, Linkage, Module};
 use cranelift_simplejit::{SimpleJITBackend, SimpleJITBuilder};
 use std::mem;
 
 fn main() {
-    let mut module: Module<SimpleJITBackend> = Module::new(SimpleJITBuilder::new(
-        SimpleJITBuilder::default_libcall_names(),
-    ));
+    let mut module: Module<SimpleJITBackend> =
+        Module::new(SimpleJITBuilder::new(default_libcall_names()));
     let mut ctx = module.make_context();
     let mut func_ctx = FunctionBuilderContext::new();
